extract parsing page to separate class parse webpaeg
